Working on android port

+++ android/jni/database/notes.cpp
@@ -1405,7 +1405,7 @@ vector <Database_Notes_Text> Database_Notes::getPossibleSeverities ()
 {
   vector <string> standard = standard_severities ();
   vector <Database_Notes_Text> severities;
-  for (unsigned int i = 0; i < standard.size(); i++) {
+  for (size_t i = 0; i < standard.size(); i++) {
     Database_Notes_Text severity;
     severity.raw = convert_to_string (i);
     severity.localized = translate (standard[i].c_str());
